@@ -4,20 +4,24 @@ dependencies {
 }
 
 android {
-    compileSdkVersion 16
-    buildToolsVersion "25.0.0"
+    compileSdk = 34
 
     defaultConfig {
-        minSdkVersion 16
-        targetSdkVersion 16
-        versionCode 2
-        versionName "1.1"
-        applicationId "org.birthdayadapter.jb.workaround"
+        minSdkVersion = 26
+        targetSdkVersion = 34
+        versionCode = 3
+        versionName = "1.2.0"
+        applicationId = "org.birthdayadapter.jb.workaround"
     }
 
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_7
-        targetCompatibility JavaVersion.VERSION_1_7
+        sourceCompatibility = JavaVersion.VERSION_11
+        targetCompatibility = JavaVersion.VERSION_11
+    }
+    namespace = 'org.birthdayadapter.jb.workaround'
+    lint {
+        abortOnError = false
+        checkReleaseBuilds = false
     }
 
     /*
@@ -36,10 +40,10 @@ android {
 
         signingConfigs {
             release {
-                storeFile file(signingStoreLocation)
-                storePassword signingStorePassword
-                keyAlias signingKeyAlias
-                keyPassword signingKeyPassword
+                storeFile = file(signingStoreLocation)
+                storePassword = signingStorePassword
+                keyAlias = signingKeyAlias
+                keyPassword = signingKeyPassword
             }
         }
 
@@ -48,11 +52,5 @@ android {
         buildTypes.release.signingConfig = null
     }
 
-    lintOptions {
-        checkReleaseBuilds false
-        // Or, if you prefer, you can continue to check for errors in release builds,
-        // but continue the build even when errors are found:
-        abortOnError false
-    }
 
 }
