@@ -1,35 +1,31 @@
 apply plugin: 'com.android.application'
+apply plugin: 'kotlin-android'
 
 dependencies {
-    compile 'com.android.support:appcompat-v7:25.0.0'
-    compile 'com.android.support:design:25.0.0'
-//    compile 'com.android.support:preference-v7:25.0.0'
-//    compile 'com.android.support:preference-v14:25.0.0'
-    compile 'com.takisoft.fix:preference-v7:25.0.0.0'
-    compile 'org.sufficientlysecure:html-textview:2.0'
-    compile 'com.pavelsikun:vintage-chroma:1.5'
+    implementation 'androidx.appcompat:appcompat:1.7.1'
+    implementation 'com.google.android.material:material:1.13.0'
+    implementation 'androidx.preference:preference:1.2.1'
+    implementation 'com.afollestad.material-dialogs:core:3.3.0'
+    implementation 'com.afollestad.material-dialogs:color:3.3.0'
+    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
+    implementation "androidx.work:work-runtime-ktx:2.11.0"
 }
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.0"
+    compileSdk = 35
 
     defaultConfig {
-        minSdkVersion 14
-        targetSdkVersion 25
-        versionCode 19
-        versionName "2.0"
+        minSdkVersion = 28
+        targetSdkVersion = 34
+        versionCode = 20
+        versionName = "2.1.0"
     }
 
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_7
-        targetCompatibility JavaVersion.VERSION_1_7
+        sourceCompatibility = JavaVersion.VERSION_11
+        targetCompatibility = JavaVersion.VERSION_11
     }
 
-    lintOptions {
-        // not everything is translated by Transifex contributors
-        disable 'MissingTranslation', 'ExtraTranslation'
-    }
 
     /*
      * To sign release build, create file gradle.properties in ~/.gradle/ with this content:
@@ -47,10 +43,10 @@ android {
 
         signingConfigs {
             release {
-                storeFile file(signingStoreLocation)
-                storePassword signingStorePassword
-                keyAlias signingKeyAlias
-                keyPassword signingKeyPassword
+                storeFile = file(signingStoreLocation)
+                storePassword = signingStorePassword
+                keyAlias = signingKeyAlias
+                keyPassword = keyPassword
             }
         }
 
@@ -59,21 +55,32 @@ android {
         buildTypes.release.signingConfig = null
     }
 
+    flavorDimensions += "version"
     productFlavors {
-        full {
+        create("full") {
+            dimension = "version"
             applicationId = "org.birthdayadapter"
-            buildConfigField "boolean", "FULL_VERSION", "true"
-            buildConfigField "boolean", "GOOGLE_PLAY_VERSION", "false"
+            buildConfigField("boolean", "FULL_VERSION", "true")
+            buildConfigField("boolean", "GOOGLE_PLAY_VERSION", "false")
         }
-        full_google {
+        create("full_google") {
+            dimension = "version"
             applicationId = "org.birthdayadapter"
-            buildConfigField "boolean", "FULL_VERSION", "true"
-            buildConfigField "boolean", "GOOGLE_PLAY_VERSION", "true"
+            buildConfigField("boolean", "FULL_VERSION", "true")
+            buildConfigField("boolean", "GOOGLE_PLAY_VERSION", "true")
         }
-        free {
+        create("free") {
+            dimension = "version"
             applicationId = "org.birthdayadapter.free"
-            buildConfigField "boolean", "FULL_VERSION", "false"
-            buildConfigField "boolean", "GOOGLE_PLAY_VERSION", "true"
+            buildConfigField("boolean", "FULL_VERSION", "false")
+            buildConfigField("boolean", "GOOGLE_PLAY_VERSION", "true")
         }
+    }
+    namespace = 'org.birthdayadapter'
+    lint {
+        disable('MissingTranslation', 'ExtraTranslation')
+    }
+    buildFeatures {
+        buildConfig = true
     }
 }
