@@ -2,7 +2,7 @@
  * Copyright (C) 2012-2016 Dominik Schürmann <dominik@dominikschuermann.de>
  *
  * This file is part of Birthday Adapter.
- * 
+ *
  * Birthday Adapter is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
@@ -31,21 +31,25 @@ import android.os.Build;
 import android.os.Bundle;
 import android.provider.CalendarContract;
 import android.provider.ContactsContract;
-import android.support.annotation.NonNull;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.preference.Preference;
-import android.support.v7.preference.PreferenceFragmentCompat;
-import android.support.v7.preference.SwitchPreferenceCompat;
+import android.view.View;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.core.content.ContextCompat;
+import androidx.preference.Preference;
+import androidx.preference.PreferenceFragmentCompat;
+import androidx.preference.SwitchPreferenceCompat;
 
 import org.birthdayadapter.R;
 import org.birthdayadapter.util.AccountHelper;
 import org.birthdayadapter.util.Constants;
 import org.birthdayadapter.util.PreferencesHelper;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Calendar;
 
 public class BasePreferenceFragment extends PreferenceFragmentCompat {
-    private BaseActivity mActivity;
     private AccountHelper mAccountHelper;
 
     private SwitchPreferenceCompat mEnabled;
@@ -54,87 +58,101 @@ public class BasePreferenceFragment extends PreferenceFragmentCompat {
 
     @Override
     public void onCreatePreferences(Bundle bundle, String s) {
-        // save prefs here
         getPreferenceManager().setSharedPreferencesName(Constants.PREFS_NAME);
         addPreferencesFromResource(R.xml.base_preferences);
     }
 
     @Override
-    public void onActivityCreated(Bundle savedInstanceState) {
-        super.onActivityCreated(savedInstanceState);
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
 
-        mActivity = (BaseActivity) getActivity();
-        mAccountHelper = new AccountHelper(mActivity, mActivity.mBackgroundStatusHandler);
-
-        // if this is the first run, enable and sync birthday adapter!
-        if (PreferencesHelper.getFirstRun(mActivity)) {
-            PreferencesHelper.setFirstRun(mActivity, false);
-
-            addAccountAndSync();
+        BaseActivity mActivity = (BaseActivity) getActivity();
+        if (mActivity == null) {
+            return;
         }
 
-        mEnabled = (SwitchPreferenceCompat) findPreference(getString(R.string.pref_enabled_key));
-        mEnabled.setOnPreferenceChangeListener(new Preference.OnPreferenceChangeListener() {
-            @Override
-            public boolean onPreferenceChange(Preference preference, Object newValue) {
-                if (newValue instanceof Boolean) {
-                    Boolean boolVal = (Boolean) newValue;
+        mAccountHelper = new AccountHelper(mActivity);
 
-                    if (boolVal) {
-                        addAccountAndSync();
-                    } else {
-                        mAccountHelper.removeAccount();
+        mEnabled = findPreference(getString(R.string.pref_enabled_key));
+        if (mEnabled != null) {
+            mEnabled.setOnPreferenceChangeListener(new Preference.OnPreferenceChangeListener() {
+                @Override
+                public boolean onPreferenceChange(Preference preference, Object newValue) {
+                    if (newValue instanceof Boolean) {
+                        if ((Boolean) newValue) {
+                            // This will trigger the permission check if needed
+                            addAccountAndSync();
+                        } else {
+                            mAccountHelper.removeAccount();
+                        }
                     }
+                    return true;
                 }
-                return true;
+            });
+        }
+
+        // On first run, set the switch to true and trigger the sync/permission flow
+        if (PreferencesHelper.getFirstRun(mActivity)) {
+            PreferencesHelper.setFirstRun(mActivity, false);
+            // We directly call addAccountAndSync which contains the permission check.
+            // We also manually set the switch to checked state.
+            if (mEnabled != null) {
+                mEnabled.setChecked(true);
             }
-        });
+        }
 
         Preference openContacts = findPreference(getString(R.string.pref_contacts_key));
-        openContacts.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-
-            @Override
-            public boolean onPreferenceClick(Preference preference) {
-                Intent intent = new Intent(Intent.ACTION_VIEW, ContactsContract.Contacts.CONTENT_URI);
-                startActivity(intent);
-                return true;
-            }
-        });
+        if (openContacts != null) {
+            openContacts.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
+                @Override
+                public boolean onPreferenceClick(Preference preference) {
+                    Intent intent = new Intent(Intent.ACTION_VIEW, ContactsContract.Contacts.CONTENT_URI);
+                    startActivity(intent);
+                    return true;
+                }
+            });
+        }
 
         Preference openCalendar = findPreference(getString(R.string.pref_calendar_key));
-        openCalendar.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-
-            @SuppressLint("NewApi")
-            @Override
-            public boolean onPreferenceClick(Preference preference) {
-                Uri.Builder builder = CalendarContract.CONTENT_URI.buildUpon();
-                builder.appendPath("time");
-                ContentUris.appendId(builder, Calendar.getInstance().getTimeInMillis());
-                Intent intent = new Intent(Intent.ACTION_VIEW)
-                        .setData(builder.build());
-                startActivity(intent);
-                return true;
-            }
-        });
-
+        if (openCalendar != null) {
+            openCalendar.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
+                @SuppressLint("NewApi")
+                @Override
+                public boolean onPreferenceClick(Preference preference) {
+                    Uri.Builder builder = CalendarContract.CONTENT_URI.buildUpon();
+                    builder.appendPath("time");
+                    ContentUris.appendId(builder, Calendar.getInstance().getTimeInMillis());
+                    Intent intent = new Intent(Intent.ACTION_VIEW).setData(builder.build());
+                    startActivity(intent);
+                    return true;
+                }
+            });
+        }
     }
 
     @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String permissions[],
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions,
                                            @NonNull int[] grantResults) {
-        switch (requestCode) {
-            case MY_PERMISSIONS_REQUEST: {
-                if (grantResults.length > 0) {
-                    for (int res : grantResults) {
-                        if (res != PackageManager.PERMISSION_GRANTED) {
-                            mEnabled.setChecked(false);
-                            return;
-                        }
+        if (requestCode == MY_PERMISSIONS_REQUEST) {
+            boolean allGranted = true;
+            if (grantResults.length > 0) {
+                for (int res : grantResults) {
+                    if (res != PackageManager.PERMISSION_GRANTED) {
+                        allGranted = false;
+                        break;
                     }
+                }
+            } else {
+                allGranted = false;
+            }
 
-                    // permission was granted
-                    mAccountHelper.addAccountAndSync();
-                    mEnabled.setChecked(true);
+            if (allGranted) {
+                // permission was granted, now we can really add the account
+                mAccountHelper.addAccountAndSync();
+            } else {
+                // permission denied, disable the feature
+                if (mEnabled != null) {
+                    mEnabled.setChecked(false);
                 }
             }
         }
@@ -148,39 +166,36 @@ public class BasePreferenceFragment extends PreferenceFragmentCompat {
 
     @TargetApi(Build.VERSION_CODES.M)
     private boolean hasPermissions() {
-        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.M) {
-            return true;
+        if (getActivity() == null) return false;
+
+        List<String> permissionsToRequest = new ArrayList<>();
+        String[] requiredPermissions = new String[]{
+                Manifest.permission.GET_ACCOUNTS,
+                Manifest.permission.READ_CONTACTS,
+                Manifest.permission.WRITE_CONTACTS,
+                Manifest.permission.READ_CALENDAR,
+                Manifest.permission.WRITE_CALENDAR
+        };
+
+        for (String permission : requiredPermissions) {
+            if (ContextCompat.checkSelfPermission(getActivity(), permission) != PackageManager.PERMISSION_GRANTED) {
+                permissionsToRequest.add(permission);
+            }
         }
 
-        // check Android 6 permission
-        int contactsPerm = ContextCompat.checkSelfPermission(getActivity(),
-                Manifest.permission.READ_CONTACTS);
-        int calendarPerm = ContextCompat.checkSelfPermission(getActivity(),
-                Manifest.permission.READ_CALENDAR);
-
-        if (contactsPerm == PackageManager.PERMISSION_GRANTED
-                && calendarPerm == PackageManager.PERMISSION_GRANTED) {
-            return true;
-        } else {
-            requestPermissions(
-                    new String[]{
-                            Manifest.permission.READ_CONTACTS,
-                            Manifest.permission.READ_CALENDAR},
-                    MY_PERMISSIONS_REQUEST);
+        if (!permissionsToRequest.isEmpty()) {
+            requestPermissions(permissionsToRequest.toArray(new String[0]), MY_PERMISSIONS_REQUEST);
             return false;
+        } else {
+            return true;
         }
     }
 
     @Override
     public void onResume() {
         super.onResume();
-
-        // If account is activated check the preference
-        if (mAccountHelper.isAccountActivated()) {
-            mEnabled.setChecked(true);
-        } else {
-            mEnabled.setChecked(false);
+        if (mAccountHelper != null && mEnabled != null) {
+            mEnabled.setChecked(mAccountHelper.isAccountActivated());
         }
     }
-
 }
