@@ -28,7 +28,7 @@ public class Constants {
     public static final String FULL_PACKAGE_NAME = "org.birthdayadapter";
 
     public static final String ACCOUNT_NAME = "Birthday Adapter";
-    public static final String ACCOUNT_TYPE = "org.birthdayadapter.account";
+    public static final String ACCOUNT_TYPE = "org.birthdayadapter.auth";
     public static final String CONTENT_AUTHORITY = "com.android.calendar";
 
     public static final Account ACCOUNT = new Account(ACCOUNT_NAME, ACCOUNT_TYPE);
